+++ CsharpExtras/ValidatedType/Numeric/Integer/BinaryDigit.cs
@@ -4,19 +4,19 @@ using System.Text;
 
 namespace CsharpExtras.ValidatedType.Numeric.Integer
 {
-    public class BinaryDigit : Validated<int>
+    public class BinaryDigit : ImmutableValidated<byte>
     {
-        public BinaryDigit(int value) : base(value)
+        public BinaryDigit(byte value) : base(value)
         {
         }
 
         protected override string ValidityConditionTextDescription => "Binary digit must be 0 or 1";
 
-        protected override bool IsValid(int t)
+        protected override bool IsValid(byte t)
         {
             return (t == 0) || (t == 1);
         }
 
-        public static explicit operator BinaryDigit(int value) => new BinaryDigit(value);
+        public static explicit operator BinaryDigit(byte value) => new BinaryDigit(value);
     }
 }